+++ after/plugin/cmp_obsidian_new.lua
@@ -1,1 +1,4 @@
-require("cmp").register_source("obsidian_new", require("cmp_obsidian_new").new())
+local has_cmp, cmp = pcall(require, "cmp")
+if has_cmp then
+  cmp.register_source("obsidian_new", require("cmp_obsidian_new").new())
+end