+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gracefully handle invalid aliases / tags in frontmatter (i.e. values that aren't strings). We'll warn about them and ignore the invalid values.
 - Fixed `nvim-cmp` completion for notes that have no `aliases` specified.
 - `nvim-cmp` completion will search based on file names now too, not just contents.
+- Fixed bug when `nvim-cmp` is not installed.
 
 ## [v1.8.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.8.0) - 2023-02-16