+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `nvim-cmp` completion for notes that have no `aliases` specified.
 - `nvim-cmp` completion will search based on file names now too, not just contents.
 - Fixed bug when `nvim-cmp` is not installed.
+- Workaround error which prevented users from using `ObsidianOpen` when vault path was configured behind a link
 
 ## [v1.8.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.8.0) - 2023-02-16